Refactor: 더미 slice 디렉토리 변경
