+++ src/shared/store/mapSlice.ts
